pyproject.toml
[project]
name = "trip-planner"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "geopy>=2.4.1",
    "google-adk>=1.15.1",
    "python-dotenv>=1.1.1",
]
